+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- @cosmjs/encoding: Create `normalizeBech32`.
 - @cosmjs/stargate: Added support for vesting messages.
 
 ## [0.28.1] - 2022-03-30